hayhooks/requirements.txt
pydantic
hayhooks
qdrant_client
ollama-haystack
psycopg[binary]
